Strip down to bare essentials
