+++ app/main.css
@@ -87,55 +87,6 @@ table th {
 #app > main.fullscreen #canvas {
   max-height: 100vh;
 }
-#app > main #menu {
-  display: block;
-  overflow: auto;
-  position: absolute;
-  border: 1px solid #fff;
-  visibility: visible;
-  background: rgba(128, 128, 128, 0.5);
-}
-#app > main #menu #banner {
-  display: block;
-  width: 50%;
-  margin: 0 auto;
-}
-#app > main #menu > p {
-  font-size: 0.8em;
-  margin: 1em auto;
-  background: #fff;
-  max-width: 40em;
-  border-radius: 1em;
-  padding: 0.5em 1em;
-}
-#app > main #menu #levels-list {
-  font-size: 0.8em;
-  margin: 1em auto;
-  background: #fff;
-  max-width: 40em;
-  border-radius: 1em;
-}
-#app > main #menu #levels-list > div {
-  padding: 0.5em 1em;
-}
-#app > main #menu #levels-list > div button {
-  cursor: pointer;
-  color: #fff;
-  background: #c44;
-  border: 2px solid #ccc;
-  border-radius: 1rem;
-  box-shadow: 0.1rem 0.1rem 0 rgba(0, 0, 0, 0.5);
-  width: 10em;
-}
-#app > main #menu #levels-list > div button:hover {
-  border: 2px solid #fff;
-}
-#app > main #menu #levels-list > div button:active {
-  border: 2px solid #888;
-}
-#app > main #menu #levels-list > div span {
-  margin-left: 1em;
-}
 #app > main .ui-button {
   position: absolute;
   cursor: pointer;
@@ -155,14 +106,6 @@ table th {
   display: block;
   width: 1.5rem;
   height: 1.5rem;
-}
-#app > main #button-home {
-  top: 0.5em;
-  left: 0.5em;
-}
-#app > main #button-fullscreen {
-  top: 0.5em;
-  right: 0.5em;
 }
 #app > main #button-reload {
   top: 3.5em;
