+++ index.html
@@ -31,11 +31,6 @@
 </header>
 <main id="main" tabindex="0">
   <canvas id="canvas"></canvas>
-  <div id="menu">
-    <div id="levels-list"></div>
-  </div>
-  <button id="button-home" class="ui-button"><img src="./assets/icon-menu-white.svg" alt="Home Menu"></button>
-  <button id="button-fullscreen" class="ui-button"><img src="./assets/icon-fullscreen-white.svg" alt="Fullscreen"></button>
   <button id="button-reload" class="ui-button"><img src="./assets/icon-refresh-white.svg" alt="Reload"></button>
 </main>
 <footer>
