+++ src/japari-bun-catch/image-asset.js
