+++ src/japari-bun-catch/index.js
@@ -1,0 +1,3 @@
+import JapariBunCatch from './japari-bun-catch'
+
+export default JapariBunCatch
