--- src/ld48/entities/wall.js
@@ -1,24 +1,0 @@
-import Entity from '../entity'
-import { SHAPES, TILE_SIZE } from '../constants'
-
-class Wall extends Entity {
-  constructor (app, col = 0, row = 0, width = 1, height = 1, cutCorner = false) {
-    super(app)
-    
-    this.colour = '#888'
-    this.solid = true
-    this.movable = false
-    this.x = col * TILE_SIZE
-    this.y = row * TILE_SIZE
-    
-    this.shape = SHAPES.POLYGON
-    this.shapePolygonPath = []
-    
-    if (cutCorner !== 'nw') this.shapePolygonPath.push(0, 0)
-    if (cutCorner !== 'ne') this.shapePolygonPath.push(width * TILE_SIZE, 0)
-    if (cutCorner !== 'se') this.shapePolygonPath.push(width * TILE_SIZE, height * TILE_SIZE)
-    if (cutCorner !== 'sw') this.shapePolygonPath.push(0, height * TILE_SIZE)
-  }
-}
-  
-export default Wall
