--- src/ld48/index.js
@@ -1,3 +1,0 @@
-import LD48 from './ld48'
-
-export default LD48
